backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
networkx>=3.2.1
python-dotenv>=1.0.0
python-multipart>=0.0.6
aiofiles>=23.2.1
httpx>=0.25.0
gunicorn